Refactoring internal functions from align to internal/alignfor.
