+++ widgets/linechart/internal/axes/label.go
@@ -21,6 +21,7 @@ import (
 	"image"
 
 	"github.com/mum4k/termdash/align"
+	"github.com/mum4k/termdash/internal/alignfor"
 )
 
 // LabelOrientation represents the orientation of text labels.
@@ -115,7 +116,7 @@ func rowLabel(scale *YScale, y int, labelWidth int) (*Label, error) {
 	}
 
 	ar := rowLabelArea(y, labelWidth)
-	pos, err := align.Text(ar, v.Text(), align.HorizontalRight, align.VerticalMiddle)
+	pos, err := alignfor.Text(ar, v.Text(), align.HorizontalRight, align.VerticalMiddle)
 	if err != nil {
 		return nil, fmt.Errorf("unable to align the label value: %v", err)
 	}
